Refactor backend architecture and fix frontend bugs
- Created `src/controllers`, `src/routes`, and `src/middlewares` to separate concerns from a bloated `server.js`.
- Implemented `express-validator` to sanitize and validate inputs on the `/api/cadastro`, `/api/login`, and `/api/clientes` routes.
- Fixed undefined variable references in `public/js/localStorage.js`.
- Consolidated frontend scripts by moving `src/js/` to `public/js/` where the static server can access them.
- Cleaned up unused and draft HTML files in `public/`.
