+++ public/js/barberpro-integration.js
