+++ public/js/chrome-compatibility.js
