+++ src/middlewares/authMiddleware.js
@@ -1,0 +1,15 @@
+const jwt = require('jsonwebtoken');
+
+function auth(req, res, next) {
+  const authHeader = req.headers.authorization;
+  if (!authHeader) return res.status(401).json({ message: 'Token não fornecido.' });
+  const token = authHeader.split(' ')[1];
+  try {
+    req.user = jwt.verify(token, process.env.JWT_SECRET);
+    next();
+  } catch {
+    res.status(401).json({ message: 'Token inválido.' });
+  }
+}
+
+module.exports = auth;
